Ensure foreman exits cleanly (#7)
Prior to this commit, the foreman process was not being killed. This was
leading to thread deadlocks.

Only the foreman process that was created during the test execution is
killed.  This is to make sure that (when running locally, for instance)
any other running foreman processes for separate apps are left intact.